Move integration tests, fix doc comment.
Integration tests for the parse module used to be included inside the
module itself. I've moved them into them into separate integration tests
where they better represent their functionality.

I've also finally removed the runner from the project, since I don't
think any of that functionality will be required well into the future.
