--- runner/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "runner"
-version.workspace = true
-edition.workspace = true
-
-[dependencies]
-axum = { version = "0.8" }
-serde.workspace = true
-tokio.workspace = true
-reqwest.workspace = true
-regex = { version = "1.10.5" }
-toml.workspace = true
-env_logger = "0.11.8"
-log = "0.4.27"
-dotenvy = "0.15.7"
-tower = "0.5.2"
-serde_json.workspace = true
