--- runner/test/example_script.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-echo "Hello from $HOSTNAME"
-for i in {0..5}; do 
-	echo "This ran $i seconds later." 
-done 
-
